workspace/pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openai>=1.93.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "quarto>=0.1.0",
    "seaborn>=0.13.2",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
]
